Separates haste tracking into its own module, adds ActiveTargets tracking, switches to EnemyInstances, adds suggestion, and cleanup
