+++ src/Parser/GuardianDruid/CombatLogParser.js
@@ -12,6 +12,7 @@ import DamageTaken from 'Parser/Core/Modules/DamageTaken';
 import Cinidaria from 'Parser/Core/Modules/Items/Cinidaria';
 
 import ActiveTargets from './Modules/Features/ActiveTargets';
+import DynamicHaste from './Modules/Features/DynamicHaste';
 import CastEfficiency from './Modules/Features/CastEfficiency';
 import Gore from './Modules/Features/Gore';
 import GalacticGuardian from './Modules/Features/GalacticGuardian';
@@ -42,8 +43,10 @@ class CombatLogParser extends CoreCombatLogParser {
     healingDone: [HealingDone, { showStatistic: true }],
     damageDone: [DamageDone, { showStatistic: true }],
     abilityTracker: AbilityTracker,
+
     // Features
     activeTargets: ActiveTargets,
+    dynamicHaste: DynamicHaste,
     castEfficiency: CastEfficiency,
     alwaysBeCasting: AlwaysBeCasting,
     goreProcs: Gore,
