+++ src/Parser/GuardianDruid/Modules/Features/ActiveTargets.js
@@ -1,7 +1,7 @@
 import Module from 'Parser/Core/Module';
 
 import Combatants from 'Parser/Core/Modules/Combatants';
-import Enemies from 'Parser/Core/Modules/Enemies';
+import { encodeTargetString } from 'Parser/Core/Modules/EnemyInstances';
 
 /*
  * The amount of time elapsed without a combat event before a target is considered inactive.
@@ -13,22 +13,23 @@ const ACTIVITY_THRESHOLD = 3000;
 class ActiveTargets extends Module {
   static dependencies = {
     combatants: Combatants,
-    enemies: Enemies,
   };
 
   /* Targets are considered inactive by default.  Once they are part of a damage event,
-   * they become active.  If they are not part of another damage event after a certain
+   * they become active.  If they are not part of another damage event before a certain
    * amount of time has passed, they are considered inactive again.
   */
   _targetActivity = {};
 
   /* Only player damage to the target is reliable for detecting enemy activity
-   * (mechanics like Shadow Shot on Sisters can trigger damage events even though
-   * the enemy is not targetable)
+   * (mechanics like Shadow Shot on Sisters can trigger enemy damage to the player events
+   * even though the enemy is not targetable).  Pet damage may be reliable too, if
+   * you were a pet class.
    */
   on_byPlayer_damage(event) {
     if (event.targetIsFriendly === false && event.targetID !== undefined) {
-      this.registerEnemyActivity(event.targetID, this.owner.currentTimestamp);
+      const enemyInstanceID = encodeTargetString(event.targetID, event.targetInstance);
+      this.registerEnemyActivity(enemyInstanceID, this.owner.currentTimestamp);
     }
   }
 
@@ -47,8 +48,10 @@ class ActiveTargets extends Module {
     const lastEvent = enemyTimeline[enemyTimeline.length - 1];
 
     if (lastEvent.end > timestamp) {
+      // Extend the current activity window
       lastEvent.end = timestamp + ACTIVITY_THRESHOLD;
     } else {
+      // Start a new activity window
       enemyTimeline.push({ start: timestamp, end: timestamp + ACTIVITY_THRESHOLD });
     }
   }
